Move mocha temp testing folder
