+++ bin/clean.js
@@ -2,6 +2,6 @@
 
 const fse = require('fs-extra');
 
-let folders = ['./built', './dist', './extension-chrome', './extension-firefox']
+let folders = ['./dist', './test/built', './extension-chrome', './extension-firefox']
 
 folders.forEach((folder) => { console.log(`Cleaning ${folder}...`); fse.removeSync(folder); });
